PR Feedback: - Split SynthesizedEmbeddedAttribute into a base class and implementation - Rename nullableFlags to NullableFlags - Rename tests
